+++ packages/api/src/StyleOptions.ts
@@ -930,12 +930,14 @@ type StyleOptions = {
   codeBlockTheme?: 'github-light-default' | 'github-dark-default';
 
   /**
-   * Feedback buttons placement
+   * (EXPERIMENTAL) Feedback buttons placement
    *
    * - `'activity-actions'` - place feedback buttons inside activity actions
    * - `'activity-status'` - place feedback buttons inside activity status
    *
    * @default 'activity-status'
+   *
+   * @deprecated This is an experimental style options and should not be used without understanding its risk.
    *
    * New in 4.19.0.
    */